backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.3.0
djoser==2.2.3
django-filter==23.1
django-extensions==3.2.3